Clean up for running in a docker container
Updated .dockerignore to ignore session file in copying.
Added requirements.txt to have an explicit list of dependencies for the project.
Updated Dockerfile to use requirements.txt when installing dependencies for the project.
Updated docker-compose.yml to reference this project as writing.com-archival.
Updated README.md to use the requirements.txt for dependencies and usage. Updated README.md to reference the container name writing.com-archival instead of scraper.